requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dictdiffer==0.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.27.1
